// frontend/src/enums/home.ts
export enum HOME_SECTION {
    DEFAULT = 'DEFAULT',
    FOR_YOU = 'FOR_YOU',
    FOLLOWING = 'FOLLOWING',
    LIKED = 'LIKED',
    SAVED = 'SAVED',
    SEARCH = 'SEARCH',
    ACTIVITY = 'ACTIVITY',
    PROFILE = 'PROFILE'
}

export enum HOME_MAIN_SECTION {
    FOR_YOU = 'FOR_YOU',
    FOLLOWING = 'FOLLOWING',
    LIKED = 'LIKED',
    SAVED = 'SAVED'
}